Boxes can be attempted to be drawn inside non-Div
The code after the if statement to prevent other boxes to be drawn in a
non-div box has been modified so that nothing the method continues as
normal, as if the if error has never been caught. Java will still
through an exception if the any objects are attempted to be drawn within
non-div boxes.